+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
-Unreleased
+Unreleased v2.0.0
 ----------
 
 This release contains breaking changes. This means that if you have been using a previous version of this library,
@@ -18,6 +18,7 @@ you _may_ need to change your code to account for the changes in this version. S
 * `getUnassignedLightByName(String)` method, to accompany the `getUnassignedLights()` method added in the previous release.
 * Added geofence support as motion sensors. They can be retrieved like all motion sensors with `getMotionSensors()` or `getMotionSensorByName(String)` with the name of a registered device.
 * Added ambient light sensor support.
+* Support for Android, thanks to the removal of the dependency to the `java.awt.Color` class (see below).
 
 ### Changed
 
@@ -25,6 +26,9 @@ you _may_ need to change your code to account for the changes in this version. S
 Instead, all switches are equal and can be accessed with the new `Switch` interface.
 The `Hue` class now has `getSwitches()` and `getSwitchByName(String)` methods instead of the old `getDimmerSwitches()` and `getDimmerSwitchByName(String)` methods.
 * `LightType.ON_OFF` enum value was renamed to `LightType.ON_OFF_LIGHT` to better distinguish it from the `ON_OFF_PLUGIN_UNIT` value.
+* Light color is to be set with the new `io.github.zeroone3010.yahueapi.Color` class. The class contains a multitude
+of factory methods, so that the transition from the old `java.awt.Color` would be as easy as possible. This change
+was made to remove the dependency on the `java.awt` package, which is not available in the Android environment.
 
 ### Removed